[mod/#31] 아이템 클릭시 id 전달

+++ app/src/main/java/com/haphap/app/presentation/search/SearchScreen.kt
@@ -56,12 +56,12 @@ private fun SearchScreen(
     uiState: SearchContract.State,
     onBackClick: () -> Unit,
     onSearchClick: () -> Unit,
-    onAutoCompleteItemClick: () -> Unit,
-    onRelatedItemClick: () -> Unit,
+    onAutoCompleteItemClick: (Int) -> Unit,
+    onRelatedItemClick: (Int) -> Unit,
     onFilterClick: (Int) -> Unit,
     onResultCardClick: (Int) -> Unit,
-    onDeleteClick: () -> Unit,
-    onTrendCardClick: () -> Unit,
+    onDeleteClick: (Int) -> Unit,
+    onTrendCardClick: (Int) -> Unit,
     modifier: Modifier = Modifier,
 ) {
     Column(
